+++ src/app/container/container.component.ts
@@ -1,7 +1,7 @@
+import { ContainerWebService } from './../shared/webservice/containerWeb.service';
 import { StateService } from './../shared/state.service';
 import { RefreshService } from './../shared/refresh.service';
 import { FormsModule } from '@angular/forms';
-import { ContainerWebService } from './../shared/containerWeb.service';
 import { PublishRequest } from './../shared/models/PublishRequest';
 import { Component, Input, OnDestroy } from '@angular/core';
 import { Router } from '@angular/router';
