+++ src/app/container/register-tool/register-tool.service.ts
@@ -1,6 +1,6 @@
+import { ContainerWebService } from './../../shared/webservice/containerWeb.service';
 import { StateService } from './../../shared/state.service';
 import { ContainerService } from './../../shared/container.service';
-import { ContainerWebService } from './../../shared/containerWeb.service';
 import { Tool } from './tool';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { Injectable, ViewChild } from '@angular/core';
