+++ src/app/shared/refresh.service.ts
@@ -1,8 +1,9 @@
-import { UsersService } from './webservice/users.service';
-import { StateService } from './state.service';
-import { ContainerService } from './container.service';
-import { ContainerWebService } from './containerWeb.service';
 import { Injectable } from '@angular/core';
+
+import { ContainerService } from './container.service';
+import { ContainerWebService } from './webservice/containerWeb.service';
+import { StateService } from './state.service';
+import { UsersService } from './webservice/users.service';
 
 @Injectable()
 export class RefreshService {
